Extract modules into separate files
